fix(resolver): move dictionaryKindOrPanic to helpers.go so gqlgen regen preserves it
gqlgen generate strips top-level helpers from schema.resolvers.go while
preserving resolver-method bodies, so defining the function inline left
the UpsertDictionary and ValidateDictionary call sites referencing an
undefined symbol after every regen. backend CI and the E2E Playwright
job both reproduced the failure on PR #125. helpers.go already documents
this constraint at its top — relocate the function (and its eris import)
there.
